pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "statusbot"
version = "0.1.0"
description = "Telegram VPS status bot with alerts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot[job-queue]==22.5",
]

[project.scripts]
statusbot = "statusbot.main:main"

[tool.hatch.build.targets.wheel]
packages = ["statusbot"]
